Adds UI Fixes and GET for fileUpload
Handlebars for sign in and up are present on page load, sign in will
bring up all other forms. Sign out reverts to page load configuration.

Adds button to view all files in the database with preliminary layout.

Adds SCSS files for team members to work on with their names for the
file name.