Adds update functionality as well as view all
After login a user now automatically will see all files in the database.
If a user fills out the update form they can update the name and tags in
the database for that particular entry. currently the user needs to
enter the mongo ID for the file they want to update, next rendition
should automatically populate this for them. next steps: add delete for
crud and the ability to GET a single item.