Adds Buttons to delete Files
"my files" now includes a delet button under each entry, button has a
data id that corresponds to the database id which will communicate with
the express API, deleting the file from both the mongo DB and also the
s3 bucket. Also adds a div to hold the results of view file instead of
appending them to the form.