scrollable file holder
added scrollfile.handlebars file to have a place to store my files and
all files. tabs are fully functional and still need to be tied in with
myfiles and allfilles which mike will be doing.
The image in the background was just something to put in there as a background
we do not need to keep it and same with the background color for the tabs.