Adds UI changes
Adds new stylesheet for my changes
Adds fix to warning for use of camelcase in file name.
Update form is now part of each of "my files", adds update form with
prepopulated ID for the selected file.
HTML for the files has been updated to add update button as well as
formatting changes to the dates.
Update handlebar has been changed to allow prepopulation of fields and
reduce title size